thanks page renders as server component
