+++ apps/course-builder-web/package.json
@@ -81,11 +81,12 @@
 		"@sanity/icons": "^2.6.0",
 		"@sindresorhus/slugify": "^2.2.1",
 		"@t3-oss/env-nextjs": "^0.7.1",
-		"@tanstack/react-query": "^4.36.1",
-		"@trpc/client": "10.45.0",
-		"@trpc/next": "10.45.0",
-		"@trpc/react-query": "10.45.0",
-		"@trpc/server": "10.45.0",
+		"@tanstack/react-query": "^5.25.0",
+		"@tanstack/react-query-devtools": "^5.31.0",
+		"@trpc/client": "next",
+		"@trpc/next": "next",
+		"@trpc/react-query": "next",
+		"@trpc/server": "next",
 		"@types/lodash": "^4.14.200",
 		"@types/mdx": "^2.0.10",
 		"@types/shortid": "^0.0.31",
