+++ apps/course-builder-web/src/app/api/trpc/[trpc]/route.ts
@@ -1,6 +1,5 @@
 import { type NextRequest } from 'next/server'
 import { env } from '@/env.mjs'
-import { withSkill } from '@/server/with-skill'
 import { appRouter } from '@/trpc/api/root'
 import { createTRPCContext } from '@/trpc/api/trpc'
 import { fetchRequestHandler } from '@trpc/server/adapters/fetch'
@@ -15,7 +14,7 @@ const createContext = async (req: NextRequest) => {
 	})
 }
 
-const baseHandler = (req: NextRequest) =>
+const handler = (req: NextRequest) =>
 	fetchRequestHandler({
 		endpoint: '/api/trpc',
 		req,
@@ -30,7 +29,5 @@ const baseHandler = (req: NextRequest) =>
 					}
 				: undefined,
 	})
-
-const handler = withSkill(baseHandler)
 
 export { handler as GET, handler as POST }
