+++ apps/course-builder-web/src/env.mjs
@@ -80,6 +80,7 @@ export const env = createEnv({
 		NEXT_PUBLIC_URL: z.string(),
 		NEXT_PUBLIC_CLOUDINARY_CLOUD_NAME: z.string(),
 		NEXT_PUBLIC_CLOUDINARY_UPLOAD_PRESET: z.string(),
+		NEXT_PUBLIC_SUPPORT_EMAIL: z.string(),
 	},
 
 	/**
@@ -104,6 +105,7 @@ export const env = createEnv({
 		INNGEST_SIGNING_KEY: process.env.INNGEST_SIGNING_KEY,
 		NEXT_PUBLIC_PARTYKIT_ROOM_NAME: process.env.NEXT_PUBLIC_PARTYKIT_ROOM_NAME,
 		NEXT_PUBLIC_PARTY_KIT_URL: process.env.NEXT_PUBLIC_PARTY_KIT_URL,
+		NEXT_PUBLIC_SUPPORT_EMAIL: process.env.NEXT_PUBLIC_SUPPORT_EMAIL,
 		MUX_ACCESS_TOKEN_ID: process.env.MUX_ACCESS_TOKEN_ID,
 		MUX_SECRET_KEY: process.env.MUX_SECRET_KEY,
 		GITHUB_CLIENT_ID: process.env.GITHUB_CLIENT_ID,
