+++ apps/course-builder-web/src/pages/thanks/redeem.tsx
@@ -1,16 +1,14 @@
 import * as React from 'react'
 import { GetServerSideProps } from 'next'
-import Layout from '@/components/app/layout'
-import { getSdk } from '@skillrecordings/database'
-
-import { LoginLink } from './purchase'
+import { courseBuilderAdapter } from '@/db'
+import { LoginLink } from '@/path-to-purchase/post-purchase-login-link'
 
 export const getServerSideProps: GetServerSideProps = async (context) => {
 	const { query } = context
 
 	const { purchaseId } = query
 
-	const { getPurchaseWithUser } = getSdk()
+	const { getPurchaseWithUser } = courseBuilderAdapter
 
 	const purchase = await getPurchaseWithUser(purchaseId as string)
 
@@ -25,7 +23,7 @@ const ThanksRedeem: React.FC<
 	React.PropsWithChildren<{ purchase: any; email: string }>
 > = ({ email }) => {
 	return (
-		<Layout footer={null}>
+		<div>
 			<main className="flex flex-grow flex-col items-center justify-center px-5 pb-16 pt-5">
 				<div className="mx-auto w-full max-w-3xl">
 					<h1 className="w-full pb-3 font-semibold uppercase tracking-wide">
@@ -34,7 +32,7 @@ const ThanksRedeem: React.FC<
 					<LoginLink email={email} />
 				</div>
 			</main>
-		</Layout>
+		</div>
 	)
 }
