+++ apps/course-builder-web/src/trpc/api/root.ts
@@ -4,7 +4,7 @@ import { pricingRouter } from '@/trpc/api/routers/pricing'
 import { purchaseUserTransferRouter } from '@/trpc/api/routers/purchaseUserTransfer'
 import { usersRouter } from '@/trpc/api/routers/users'
 import { videoResourceRouter } from '@/trpc/api/routers/videoResource'
-import { createTRPCRouter } from '@/trpc/api/trpc'
+import { createCallerFactory, createTRPCRouter } from '@/trpc/api/trpc'
 
 /**
  * This is the primary router for your server.
@@ -22,3 +22,12 @@ export const appRouter = createTRPCRouter({
 
 // export type definition of API
 export type AppRouter = typeof appRouter
+
+/**
+ * Create a server-side caller for the tRPC API.
+ * @example
+ * const trpc = createCaller(createContext);
+ * const res = await trpc.post.all();
+ *       ^? Post[]
+ */
+export const createCaller = createCallerFactory(appRouter)
