+++ apps/course-builder-web/src/trpc/api/routers/ability.ts
@@ -4,7 +4,9 @@ import { createTRPCRouter, publicProcedure } from '@/trpc/api/trpc'
 
 export const abilityRouter = createTRPCRouter({
 	getCurrentAbilityRules: publicProcedure.query(async () => {
-		const { session } = await getServerAuthSession()
-		return getAbilityRules({ user: session?.user })
+		// const { session } = await getServerAuthSession()
+		// return getAbilityRules({ user: session?.user })
+
+		return getAbilityRules({})
 	}),
 })
