+++ apps/course-builder-web/src/trpc/api/routers/users.ts
@@ -1,9 +1,13 @@
 import { db } from '@/db'
+import { accounts } from '@/db/schema'
+import { getServerAuthSession } from '@/server/auth'
 import {
 	createTRPCRouter,
 	protectedProcedure,
 	publicProcedure,
 } from '@/trpc/api/trpc'
+import { and, eq } from 'drizzle-orm'
+import { isEmpty } from 'lodash'
 import { z } from 'zod'
 
 export const usersRouter = createTRPCRouter({
@@ -18,4 +22,18 @@ export const usersRouter = createTRPCRouter({
 				where: (users, { eq }) => eq(users.id, input.userId),
 			})
 		}),
+
+	githubConnected: publicProcedure.query(async ({ ctx, input }) => {
+		const token = await getServerAuthSession()
+		if (!token.session?.user) return false
+
+		const userAccounts = await db.query.accounts.findMany({
+			where: and(
+				eq(accounts.userId, token.session.user.id),
+				eq(accounts.provider, 'github'),
+			),
+		})
+
+		return !isEmpty(userAccounts)
+	}),
 })
