+++ apps/course-builder-web/src/trpc/api/trpc.ts
@@ -58,6 +58,13 @@ const t = initTRPC.context<typeof createTRPCContext>().create({
 })
 
 /**
+ * Create a server-side caller.
+ *
+ * @see https://trpc.io/docs/server/server-side-calls
+ */
+export const createCallerFactory = t.createCallerFactory
+
+/**
  * 3. ROUTER & PROCEDURE (THE IMPORTANT BIT)
  *
  * These are the pieces you use to build your tRPC API. You should import these a lot in the
