+++ apps/course-builder-web/src/trpc/react.tsx
@@ -1,36 +1,32 @@
 'use client'
 
 import { useState } from 'react'
-import { AppRouter } from '@/trpc/api/root'
+import { type AppRouter } from '@/trpc/api/root'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import { loggerLink, unstable_httpBatchStreamLink } from '@trpc/client'
 import { createTRPCReact } from '@trpc/react-query'
+import SuperJSON from 'superjson'
 
-import { getUrl, transformer } from './shared'
+const createQueryClient = () => new QueryClient()
 
-export const api = createTRPCReact<AppRouter>({
-	overrides: {
-		useMutation: {
-			async onSuccess(opts) {
-				/**
-				 * @note that order here matters: The order here allows route changes in `onSuccess` without
-				 *       having a flash of content change whilst redirecting.
-				 */
-				// Calls the `onSuccess` defined in the `useQuery()`-options:
-				await opts.originalFn()
-				// Invalidate all queries in the react-query cache:
-				await opts.queryClient.invalidateQueries()
-			},
-		},
-	},
-})
+let clientQueryClientSingleton: QueryClient | undefined = undefined
+const getQueryClient = () => {
+	if (typeof window === 'undefined') {
+		// Server: always make a new query client
+		return createQueryClient()
+	}
+	// Browser: use singleton pattern to keep the same query client
+	return (clientQueryClientSingleton ??= createQueryClient())
+}
+
+export const api = createTRPCReact<AppRouter>()
 
 export function TRPCReactProvider(props: { children: React.ReactNode }) {
-	const [queryClient] = useState(() => new QueryClient())
+	const queryClient = getQueryClient()
 
 	const [trpcClient] = useState(() =>
 		api.createClient({
-			transformer,
 			links: [
 				loggerLink({
 					enabled: (op) =>
@@ -38,7 +34,13 @@ export function TRPCReactProvider(props: { children: React.ReactNode }) {
 						(op.direction === 'down' && op.result instanceof Error),
 				}),
 				unstable_httpBatchStreamLink({
-					url: getUrl(),
+					transformer: SuperJSON,
+					url: getBaseUrl() + '/api/trpc',
+					headers: () => {
+						const headers = new Headers()
+						headers.set('x-trpc-source', 'nextjs-react')
+						return headers
+					},
 				}),
 			],
 		}),
@@ -49,6 +51,13 @@ export function TRPCReactProvider(props: { children: React.ReactNode }) {
 			<api.Provider client={trpcClient} queryClient={queryClient}>
 				{props.children}
 			</api.Provider>
+			<ReactQueryDevtools initialIsOpen={false} />
 		</QueryClientProvider>
 	)
+}
+
+function getBaseUrl() {
+	if (typeof window !== 'undefined') return window.location.origin
+	if (process.env.VERCEL_URL) return `https://${process.env.VERCEL_URL}`
+	return `http://localhost:${process.env.PORT ?? 3000}`
 }
