+++ cli/package.json
@@ -69,11 +69,11 @@
 		"@planetscale/database": "1.14.0",
 		"@prisma/client": "^5.6.0",
 		"@t3-oss/env-nextjs": "^0.7.1",
-		"@tanstack/react-query": "^4.36.1",
-		"@trpc/client": "10.45.0",
-		"@trpc/next": "10.45.0",
-		"@trpc/react-query": "10.45.0",
-		"@trpc/server": "10.45.0",
+		"@tanstack/react-query": "^5.25.0",
+		"@trpc/client": "next",
+		"@trpc/next": "next",
+		"@trpc/react-query": "next",
+		"@trpc/server": "next",
 		"@types/fs-extra": "^11.0.1",
 		"@types/gradient-string": "^1.1.2",
 		"@types/inquirer": "^9.0.3",
