+++ packages/core/src/schemas/product-schema.ts
@@ -1,5 +1,6 @@
 import { z } from 'zod'
 
+import { ContentResourceSchema } from './content-resource-schema'
 import { priceSchema } from './price-schema'
 
 export const productSchema = z.object({
@@ -11,6 +12,7 @@ export const productSchema = z.object({
 	status: z.number().int().default(0),
 	quantityAvailable: z.number().int().default(-1),
 	price: priceSchema.nullable().optional(),
+	resources: z.array(ContentResourceSchema).default([]).nullable(),
 })
 
 export type Product = z.infer<typeof productSchema>
