+++ packages/core/src/schemas/user-schema.ts
@@ -1,12 +1,12 @@
 import { z } from 'zod'
 
 export const userSchema = z.object({
-	id: z.string().length(255),
-	name: z.string().length(255).optional().nullable(),
+	id: z.string().max(255),
+	name: z.string().max(255).optional().nullable(),
 	role: z.enum(['user', 'admin']).default('user'),
-	email: z.string().length(255).email(),
+	email: z.string().max(255).email(),
 	emailVerified: z.date().nullable(),
-	image: z.string().length(255).optional().nullable(),
+	image: z.string().max(255).optional().nullable(),
 	createdAt: z.date().nullable(),
 })
